Forgot to explicitly cast the vector size to int.
svn path=/trunk/koffice/; revision=751629

+++ krita/plugins/painterlyframework/kis_illuminant_profile.cpp
@@ -127,7 +127,7 @@ bool KisIlluminantProfile::save(const QString &fileName)
         for (int j = 0; j < (int)m_T->size2; j++)
             data << gsl_matrix_get(m_T, i, j);
 
-    data << m_P->size;
+    data << (int)m_P->size;
     for (int i = 0; i < (int)m_P->size; i++)
         data << gsl_vector_get(m_P, i);
